+++ controllers/like_controller.py
@@ -6,6 +6,7 @@ from flask_jwt_extended import jwt_required, get_jwt_identity
 from init import db
 from models.like import Like, like_schema, likes_schema
 from models.post import Post 
+from utils import authorise_as_admin
 
 # a like cannot exist without a post, it belongs to a post, so we can
 # register Blueprint to posts_bp (posts blueprint), and therefore it will
@@ -16,7 +17,20 @@ likes_bp = Blueprint("likes", __name__, url_prefix="/<int:post_id>/likes")
 # no need to create a fetch all likes route because it would have no purpose,
 # we only want all the likes linked to one post, which we get when fetching posts.
 
+# post/<int:post_id>/likes - GET - fetch all likes on a post
+@likes_bp.route("/")
+@jwt_required()
+def fetch_all_likes_on_post(post_id):
+    # fetch the post with the correct id - post_id (passed in url)
+    stmt = db.select(Like).filter_by(post_id=post_id)
+    likes = db.session.scalars(stmt)
 
+    if likes: 
+        
+        return likes_schema.dump(likes)
+    
+    else:
+        return {"error": f"Post with id {post_id} not found"}, 404
 
 # Create like route
 @likes_bp.route("/", methods=["POST"])
@@ -57,6 +71,11 @@ def delete_like(post_id, like_id):
     like = db.session.scalar(stmt)
     # if like exists
     if like:
+        # check whether the user is an admin 
+        is_admin = authorise_as_admin()
+        # if the user is not the owner of the post
+        if not is_admin and str(like.user_id) != get_jwt_identity():
+            return {"error": "User unorthorised to perform this request"}, 403
         # delete like
         db.session.delete(like)
         db.session.commit()
